docker/requirements.txt
apache-airflow==2.6.1
google-cloud-storage==2.9.0
google-cloud-bigquery==3.10.0
google-auth==2.22.0
pyarrow==12.0.0
dbt-bigquery==1.5.0 